Move TFIDF to remote
