+++ dsets/__init__.py
@@ -2,3 +2,5 @@ from .attr_snippets import AttributeSnippets
 from .counterfact import CounterFactDataset
 from .knowns import KnownsDataset
 from .zsre import MENDQADataset
+
+from .tfidf_stats import get_tfidf_vectorizer
